requirements.txt
PyQt5==5.15.9
mysql-connector-python
python-dotenv
